+++ src/main/scala/com/builtbroken/mc/prefab/explosive/Explosive.java
@@ -71,7 +71,7 @@ public class Explosive implements IExplosive
     }
 
     @Override
-    public IWorldChangeAction createBlastForTrigger(World world, double x, double y, double z, TriggerCause triggerCause, int yieldMultiplier, NBTTagCompound tag)
+    public IWorldChangeAction createBlastForTrigger(World world, double x, double y, double z, TriggerCause triggerCause, double yieldMultiplier, NBTTagCompound tag)
     {
         try
         {
